requirements.txt
accelerate==0.33.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
decorator==4.4.2
dtw-python==1.5.1
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==24.1
pillow==10.4.0
proglog==0.1.10
psutil==6.0.0
pycparser==2.22
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scipy==1.14.1
soundfile==0.12.1
sympy==1.13.2
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.5
transformers==4.44.2
triton==2.3.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
whisper-timestamped==1.15.4
